Updated project name web automation selenium test.
