+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/Models/SeleniumEasy/CheckboxDemo.cs
@@ -1,4 +1,4 @@
-﻿namespace SeleniumTests.SeleniumEasy.Models
+﻿namespace WebAutomationSeleniumTests.SeleniumEasy.Models
 {
     sealed class CheckboxDemo
     {
