+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/PageObjects/SeleniumEasy/JQuerySelectDropdown.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SeleniumTests.SeleniumEasy.Models
+namespace WebAutomationSeleniumTests.SeleniumEasy.PageObjects
 {
     class JQuerySelectDropdown
     {
