+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/PageObjects/SeleniumEasy/SelectDropdownList.cs
@@ -1,9 +1,9 @@
 ﻿using OpenQA.Selenium;
 using OpenQA.Selenium.Support.UI;
-using SeleniumTests.SeleniumEasy.Common;
+using WebAutomationSeleniumTests.SeleniumEasy.Common;
 using System.Collections.Generic;
 
-namespace SeleniumTests.SeleniumEasy.PageObjects
+namespace WebAutomationSeleniumTests.SeleniumEasy.PageObjects
 {
     sealed class SelectDropdownList
     {
