+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SeleniumTests.SeleniumEasy")]
+[assembly: AssemblyTitle("WebAutomationSeleniumTests.SeleniumEasy")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SeleniumTests.SeleniumEasy")]
+[assembly: AssemblyProduct("WebAutomationSeleniumTests.SeleniumEasy")]
 [assembly: AssemblyCopyright("Copyright ©  2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
