+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/Tests/SeleniumEasy/InputForm.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SeleniumTests.SeleniumEasy.Tests
+namespace WebAutomationSeleniumTests.SeleniumEasy.Tests
 {
-    class RadioButtons
+    class InputForm
     {
     }
 }
