+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/WebAutomationSeleniumTests.csproj
@@ -60,40 +60,40 @@
   <ItemGroup>
     <Compile Include="Common\IXpathReader.cs" />
     <Compile Include="Common\XpathReader.cs" />
-    <Compile Include="Models\AjaxFormSubmit.cs" />
-    <Compile Include="Models\CheckboxDemo.cs" />
-    <Compile Include="Models\HomePage.cs" />
-    <Compile Include="Models\InputFormSubmit.cs" />
-    <Compile Include="Models\JQuerySelectDropdown.cs" />
-    <Compile Include="Models\RadioButtonsDemo.cs" />
-    <Compile Include="Models\SelectDropdownList.cs" />
-    <Compile Include="Models\SimpleFormDemo.cs" />
-    <Compile Include="PageObjects\AjaxFormSubmit.cs" />
-    <Compile Include="PageObjects\CheckboxDemo.cs" />
-    <Compile Include="PageObjects\HomePage.cs" />
-    <Compile Include="PageObjects\InputFormSubmit.cs" />
-    <Compile Include="PageObjects\JQuerySelectDropdown.cs" />
-    <Compile Include="PageObjects\RadioButtonsDemo.cs" />
-    <Compile Include="PageObjects\SelectDropdownList.cs" />
-    <Compile Include="PageObjects\SimpleFormDemo.cs" />
+    <Compile Include="Models\SeleniumEasy\AjaxFormSubmit.cs" />
+    <Compile Include="Models\SeleniumEasy\CheckboxDemo.cs" />
+    <Compile Include="Models\SeleniumEasy\HomePage.cs" />
+    <Compile Include="Models\SeleniumEasy\InputFormSubmit.cs" />
+    <Compile Include="Models\SeleniumEasy\JQuerySelectDropdown.cs" />
+    <Compile Include="Models\SeleniumEasy\RadioButtonsDemo.cs" />
+    <Compile Include="Models\SeleniumEasy\SelectDropdownList.cs" />
+    <Compile Include="Models\SeleniumEasy\SimpleFormDemo.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\AjaxFormSubmit.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\CheckboxDemo.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\HomePage.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\InputFormSubmit.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\JQuerySelectDropdown.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\RadioButtonsDemo.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\SelectDropdownList.cs" />
+    <Compile Include="PageObjects\SeleniumEasy\SimpleFormDemo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tests\AjaxForm.cs" />
-    <Compile Include="Tests\Checkbox.cs" />
-    <Compile Include="Tests\DropdownList.cs" />
-    <Compile Include="Tests\InputForm.cs" />
-    <Compile Include="Tests\JQueryDropdown.cs" />
-    <Compile Include="Tests\RadioButtons.cs" />
-    <Compile Include="Tests\SimpleForm.cs" />
+    <Compile Include="Tests\SeleniumEasy\AjaxForm.cs" />
+    <Compile Include="Tests\SeleniumEasy\Checkbox.cs" />
+    <Compile Include="Tests\SeleniumEasy\DropdownList.cs" />
+    <Compile Include="Tests\SeleniumEasy\InputForm.cs" />
+    <Compile Include="Tests\SeleniumEasy\JQueryDropdown.cs" />
+    <Compile Include="Tests\SeleniumEasy\RadioButtons.cs" />
+    <Compile Include="Tests\SeleniumEasy\SimpleForm.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-    <None Include="Xpaths\AjaxFormSubmit.json" />
-    <None Include="Xpaths\CheckboxDemo.json" />
-    <None Include="Xpaths\HomePage.json" />
-    <None Include="Xpaths\InputFormSubmit.json" />
-    <None Include="Xpaths\RadioButtonsDemo.json" />
-    <None Include="Xpaths\SelectDropdownList.json" />
-    <None Include="Xpaths\SimpleFormDemo.json" />
+    <None Include="Xpaths\SeleniumEasy\AjaxFormSubmit.json" />
+    <None Include="Xpaths\SeleniumEasy\CheckboxDemo.json" />
+    <None Include="Xpaths\SeleniumEasy\HomePage.json" />
+    <None Include="Xpaths\SeleniumEasy\InputFormSubmit.json" />
+    <None Include="Xpaths\SeleniumEasy\RadioButtonsDemo.json" />
+    <None Include="Xpaths\SeleniumEasy\SelectDropdownList.json" />
+    <None Include="Xpaths\SeleniumEasy\SimpleFormDemo.json" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
