+++ src/SeleniumTests/SeleniumTests.SeleniumEasy/Xpaths/SeleniumEasy/CheckboxDemo.json
