+++ src/SeleniumTests/SeleniumTests.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30225.117
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeleniumTests.SeleniumEasy", "SeleniumTests.SeleniumEasy\SeleniumTests.SeleniumEasy.csproj", "{BB96865C-0BCC-4F61-9450-8011855783F3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebAutomationSeleniumTests", "SeleniumTests.SeleniumEasy\WebAutomationSeleniumTests.csproj", "{BB96865C-0BCC-4F61-9450-8011855783F3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
